FLAS-40-Cards refactor: Cards refactor
