--- src/pages/mockPack/Pack/index.ts
@@ -1,1 +1,0 @@
-export * from './cards'
