--- src/pages/mockPack/index.ts
@@ -1,1 +1,0 @@
-export * from './mockPack'
