--- src/pages/mockPack/mockPack.module.scss
@@ -1,9 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  row-gap: 40px;
-  align-items: center;
-  justify-content: center;
-
-  width: 100%;
-}
