--- src/pages/mockPack/mockPack.tsx
@@ -1,7 +1,0 @@
-import s from './mockPack.module.scss'
-
-type Props = {}
-
-export const MockPack = ({}: Props) => {
-  return <section className={s.wrapper}>Cards</section>
-}
